requirements.txt
python-dotenv
phonenumbers
opencnam